Removed interface HarvestResultManager
